requirements.txt
python-telegram-bot[ext]
python-dotenv
aiohttp
selenium
playwright
webdriver-manager